cave-fire-proposals/requirements.txt
fastapi[all]
uvicorn[standard]
python-docx
Jinja2
requests
pytest
